requirements.txt
flask
pillow
opencv-python
numpy
torch
torchvision
git+https://github.com/facebookresearch/detectron2.git

